Add draw_from_bag logic when tiles drawn is greater than # of tiles in bag with tests
Add draw_from_bag logic to refill the bag when the number drawn is greater than the
number available.
Add tests for confirm the refill from tower works as part draw_from_bag.
Add test to confirm only the minimum number of tiles are transferred from bag to
destination when the combined quantity of tiles in the bag and tower is less than
the tiles needed.